Added component TagType.jsx
