+++ frontend/src/components/PokedexCard.jsx
@@ -1,20 +1,8 @@
-import { useEffect, useState } from "react";
 import PropTypes from "prop-types";
 import colors from "./colors";
-import iconType from "./iconType";
+import TagType from "./TagType";
 
 function PokedexCard({ pokemon }) {
-  const [typeImage, setTypeImage] = useState("");
-  const loadImage = () => {
-    import(iconType[pokemon.getFirstType()]).then((image) => {
-      setTypeImage(image.default);
-    });
-  };
-
-  useEffect(() => {
-    loadImage();
-  }, []);
-
   return (
     <div id="pokedexCard">
       <div className="cardPokedex">
@@ -31,13 +19,7 @@ function PokedexCard({ pokemon }) {
         </li>
         <div id="name_type">
           <li className="liPokedex">{pokemon.name}</li>
-          <li
-            className="liPokedexType"
-            style={{ backgroundColor: colors[pokemon.getFirstType()] }}
-          >
-            <div>{pokemon.getFirstType()}</div>
-            <img className="liPokedexTypeImage" src={typeImage} alt="" />
-          </li>
+          <TagType type={pokemon.getFirstType()} />
         </div>
       </ul>
     </div>
