--- frontend/src/components/SearchBar.jsx
@@ -1,23 +1,0 @@
-// import { useState } from "react";
-// // import PropTypes from "prop-types";
-
-// function SearchBar() {
-//   const [filter, setFilter] = useState("");
-
-//   const handleSearchChange = (e) => {
-//     setFilter(e.target.value);
-//   };
-
-//   return (
-//     <div>
-//       <input
-//         className="searchBarPok"
-//         type="text"
-//         onChange={handleSearchChange}
-//         // onKeyDown={(e) => setText(e.target.value)}
-//         placeholder="Pokemon"
-//       />
-//     </div>
-//   );
-// }
-// export default SearchBar;
